refactor(trusty-installer): move the pinned test suite into pinned/tests.rs
The fixture-server suite is ~550 lines and was inline under
`#[cfg(all(test, unix))]`, which counts toward the 500-SLOC production cap —
only the bare `#[cfg(test)] mod x { … }` shape is excluded (#5153), so
pinned.rs measured 756 SLOC and tripped check_line_cap.sh.

`pinned/tests.rs` is classified a test file by basename, so it gets the 3000
cap. Also repoints three `Test:` citations at their real test names.
